Make Functions Async
