+++ Basecode.Services/Interfaces/IInterviewsService.cs
@@ -17,8 +17,9 @@ namespace Basecode.Services.Interfaces
         Task<Interviews?> GetByIdAsync(int id);
         Task UpdateAsync(InterviewsUpdationDto Interviews);
         Task DeleteAsync(int id);
-        IEnumerable<Interviews> GetInterviewsByInterviewerAndDate(int interviewerId, DateTime interviewDate);
-        IEnumerable<Interviews> GetInterviewsByInterviewer(int interviewerId);
-        IEnumerable<Interviews> GetInterviewsByApplicant(int applicantId);
+        Task<IEnumerable<Interviews>> GetInterviewsByInterviewerAndDateAsync(int interviewerId, DateTime interviewDate);
+        Task<IEnumerable<Interviews>> GetInterviewsByInterviewerAsync(int interviewerId);
+        Task<IEnumerable<Interviews>> GetInterviewsByApplicantAsync(int applicantId);
+        Task<bool> IsTimeRangeOverlapping(InterviewsCreationDto newInterview);
     }
 }
