+++ Basecode.Services/Services/InterviewsService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -60,34 +61,87 @@ namespace Basecode.Services.Services
             return await _repository.GetByApplicantIdAsync(applicantId);
         }
 
-        public IEnumerable<Interviews> GetInterviewsByInterviewerAndDate(int interviewerId, DateTime interviewDate)
+        public async Task<IEnumerable<Interviews>> GetInterviewsByInterviewerAndDateAsync(int interviewerId, DateTime interviewDate)
         {
-            return _repository.GetInterviewsByInterviewerAndDate(interviewerId, interviewDate);
+            return await _repository.GetInterviewsByInterviewerAndDateAsync(interviewerId, interviewDate);
         }
 
-        public IEnumerable<Interviews> GetInterviewsByInterviewer(int interviewerId)
+        public async Task<IEnumerable<Interviews>> GetInterviewsByInterviewerAsync(int interviewerId)
         {
-            return _repository.GetInterviewsByInterviewer(interviewerId);
+            return await _repository.GetInterviewsByInterviewerAsync(interviewerId);
         }
 
-        public IEnumerable<Interviews> GetInterviewsByApplicant(int applicantId)
+        public async Task<IEnumerable<Interviews>> GetInterviewsByApplicantAsync(int applicantId)
         {
-            return _repository.GetInterviewsByApplicant(applicantId);
+            return await _repository.GetInterviewsByApplicantAsync(applicantId);
         }
 
-        public IEnumerable<Interviews> GetInterviewsByInterviewerAndDate(int interviewerId, DateTime interviewDate)
+        public async Task<bool> IsTimeRangeOverlappingAsync(InterviewsCreationDto newInterview)
         {
-            return _repository.GetInterviewsByInterviewerAndDate(interviewerId, interviewDate);
-        }
+            var existingInterviews = await GetInterviewsByInterviewerAndDateAsync(newInterview.InterviewerId, newInterview.InterviewDate);
 
-        public IEnumerable<Interviews> GetInterviewsByInterviewer(int interviewerId)
-        {
-            return _repository.GetInterviewsByInterviewer(interviewerId);
-        }
+            var newStart = DateTime.ParseExact(newInterview.TimeStart, "h:mm tt", CultureInfo.InvariantCulture);
+            var newEnd = DateTime.ParseExact(newInterview.TimeEnd, "h:mm tt", CultureInfo.InvariantCulture);
 
-        public IEnumerable<Interviews> GetInterviewsByApplicant(int applicantId)
-        {
-            return _repository.GetInterviewsByApplicant(applicantId);
+            foreach (var existingInterview in existingInterviews)
+            {
+                var existingStart = DateTime.ParseExact(existingInterview.TimeStart, "h:mm tt", CultureInfo.InvariantCulture);
+                var existingEnd = DateTime.ParseExact(existingInterview.TimeEnd, "h:mm tt", CultureInfo.InvariantCulture);
+
+                if (newInterview.InterviewDate == existingInterview.InterviewDate)
+                {
+                    if ((newStart >= existingStart && newStart < existingEnd) ||
+                        (newEnd > existingStart && newEnd <= existingEnd) ||
+                        (newStart <= existingStart && newEnd >= existingEnd))
+                    {
+                        return true;
+                    }
+                }
+            }
+
+            var otherInterviewsByInterviewer = await GetInterviewsByInterviewerAsync(newInterview.InterviewerId);
+            foreach (var otherInterview in otherInterviewsByInterviewer)
+            {
+                if (otherInterview.Id != newInterview.Id)
+                {
+                    var otherStart = DateTime.ParseExact(otherInterview.TimeStart, "h:mm tt", CultureInfo.InvariantCulture);
+                    var otherEnd = DateTime.ParseExact(otherInterview.TimeEnd, "h:mm tt", CultureInfo.InvariantCulture);
+
+                    if (newInterview.InterviewDate == otherInterview.InterviewDate)
+                    {
+                        if ((newStart >= otherStart && newStart < otherEnd) ||
+                            (newEnd > otherStart && newEnd <= otherEnd) ||
+                            (newStart <= otherStart && newEnd >= otherEnd))
+                        {
+                            Console.WriteLine(2);
+                            return true;
+                        }
+                    }
+                }
+            }
+
+            var otherInterviewsByApplicant = await GetInterviewsByApplicantAsync(newInterview.ApplicantId);
+            foreach (var otherInterview in otherInterviewsByApplicant)
+            {
+                if (otherInterview.Id != newInterview.Id)
+                {
+                    var otherStart = DateTime.ParseExact(otherInterview.TimeStart, "h:mm tt", CultureInfo.InvariantCulture);
+                    var otherEnd = DateTime.ParseExact(otherInterview.TimeEnd, "h:mm tt", CultureInfo.InvariantCulture);
+
+                    if (newInterview.InterviewDate == otherInterview.InterviewDate)
+                    {
+                        if ((newStart >= otherStart && newStart < otherEnd) ||
+                            (newEnd > otherStart && newEnd <= otherEnd) ||
+                            (newStart <= otherStart && newEnd >= otherEnd))
+                        {
+                            Console.WriteLine(3);
+                            return true;
+                        }
+                    }
+                }
+            }
+
+            return false;
         }
 
     }
